code is cleaned and ready to use depencency injection pattern
