+++ Task2/Kuba/NoughtsAndCrosses/Game.cs
@@ -4,45 +4,36 @@ namespace NoughtsAndCrosses
 {
     class Game
     {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Welcome to Nougths And Crosses!");
-            Player player1 = new Player();
-            Player player2 = new Player();
-            Console.WriteLine("Player 1 will play with noughts. Choose player 1 name:");
-            player1.Name = Console.ReadLine();
-            player1.ChoosenSymbol = GameBoardStatus.nought;
-            Console.WriteLine("Player 2 will play with crosses. Choose player 2 name:");
-            player2.Name = Console.ReadLine();
-            player2.ChoosenSymbol = GameBoardStatus.cross;
-
-            GameBoard board = new GameBoard();
-            GameLoop(player1, player2, board);
-
-            Console.ReadLine();
-        }
-
-        private static void GameLoop(Player player1, Player player2, GameBoard board)
+        public void GameLoop(Player player1, Player player2, GameBoard board)
         {
             BoardConsoleTranslator consoleInterpreter = new BoardConsoleTranslator();
             bool gameOver = false;
+            Player currentPlayer = player1;
             while (!gameOver)
             {
-                MakeMove(board, player1);
-                gameOver = GameIsOver(board, player1);
+                MakeMove(board, currentPlayer);
+                gameOver = GameIsOver(board, currentPlayer);
                 consoleInterpreter.PrintToConsole(board);
                 if (gameOver) break;
+                currentPlayer = ChangePlayer(currentPlayer, player1, player2);
+            }
+        }
 
-                MakeMove(board, player2);
-                gameOver = GameIsOver(board, player2);
-                consoleInterpreter.PrintToConsole(board);
-                if (gameOver) break;
+        private static Player ChangePlayer(Player currentPlayer, Player player1, Player player2)
+        {
+            if (currentPlayer == player1)
+            {
+                return player2;
+            }
+            else
+            {
+                return player1;
             }
         }
 
         private static void MakeMove(GameBoard board, Player player)
         {
-            Position coordinates = new Position();
+            Position coordinates;
             BoardConsoleTranslator consoleInterpreter = new BoardConsoleTranslator();
             coordinates = consoleInterpreter.AskForCoordinatesUntilValid(board);
             board.MarkPlayerTurn(player, coordinates);
